+++ worlds/banjo_tooie/BTClient.py
@@ -270,6 +270,7 @@ def get_slot_payload(ctx: BanjoTooieContext):
             "slot_world_order": ctx.slot_data["world_order"],
             "slot_keys": ctx.slot_data['world_keys'],
             "slot_skip_klungo": ctx.slot_data["skip_klungo"],
+            "slot_skip_king": ctx.slot_data["skip_king"],
             "slot_goal_type": ctx.slot_data["goal_type"],
             "slot_minigame_hunt_length": ctx.slot_data["minigame_hunt_length"],
             "slot_boss_hunt_length": ctx.slot_data["boss_hunt_length"],